cleaned up pages to add components for each page
