+++ src/components/QuizContent/QuizContent.scss
@@ -1,0 +1,73 @@
+.quiz-page {
+  max-width: 640px;
+  margin: 9rem auto 4rem;
+  padding: 3rem 2rem;
+  border-radius: 1.5rem;
+  font-family: "Garbata", serif;
+  color: #5a2d2f;
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center; // Center content horizontally
+  justify-content: center; // Center content vertically
+  min-height: 70vh; // Ensure the container takes up enough space
+
+  &__content {
+    width: 100%;
+    text-align: center;
+    padding: 2rem;
+  }
+
+  &__question {
+    scroll-margin-top: 100px;
+    padding: 1rem;
+    border-radius: 1rem;
+    margin-bottom: 6rem;
+    animation: slideFadeIn 0.5s ease forwards;
+    opacity: 0;
+    transform: translateY(30px);
+
+    h3 {
+      font-size: 1.25rem;
+      font-weight: 600;
+      margin-bottom: 1rem;
+      text-align: center;
+    }
+  }
+
+  &__options {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    align-items: center;
+  }
+  &__submit-button {
+    margin-top: 1.5rem;
+    width: 100%;
+  }
+
+  &__option-button,
+  &__submit-button {
+    background-color: #83494c;
+    color: #fff0eb;
+    padding: 0.85rem 2rem;
+    border: none;
+    border-radius: 2rem;
+    font-family: inherit;
+    font-size: 1.125rem;
+    font-weight: 600;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    box-shadow: 0 4px 12px rgba(131, 73, 76, 0.2);
+
+    &:hover {
+      background-color: #6a3a3d;
+      transform: scale(1.05);
+    }
+  }
+
+  &__submit-button {
+    margin-top: 1.5rem;
+    width: 100%;
+  }
+}
