+++ src/components/RegisterForm/RegisterForm.scss
@@ -1,0 +1,104 @@
+.register-page {
+  max-width: 640px;
+  margin: 9rem auto 4rem;
+  padding: 3rem 2rem;
+  border-radius: 1.5rem;
+  font-family: "Garbata", serif;
+  color: #5a2d2f;
+  position: relative;
+
+  &__title {
+    font-size: 2.75rem;
+    font-weight: 600;
+    text-align: center;
+    margin-bottom: 2rem;
+    color: #83494c;
+    text-transform: uppercase;
+    letter-spacing: 0.5px;
+  }
+
+  &__error {
+    color: #e63946;
+    background-color: #fff3f3;
+    padding: 0.75rem 1rem;
+    border-left: 5px solid #e63946;
+    border-radius: 0.5rem;
+    font-weight: 500;
+    margin-bottom: 1.5rem;
+    text-align: center;
+  }
+
+  &__form {
+    display: flex;
+    flex-direction: column;
+    gap: 1.75rem;
+  }
+
+  &__form-group {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+
+    label {
+      font-size: 1rem;
+      font-weight: 600;
+      color: #6a3a3d;
+    }
+
+    input,
+    select {
+      padding: 0.85rem 1rem;
+      border: 2px solid #83494c;
+      border-radius: 0.5rem;
+      font-family: inherit;
+      font-size: 1rem;
+      color: #5a2d2f;
+      background-color: #fff;
+      transition: border-color 0.3s ease, box-shadow 0.3s ease;
+
+      &:focus {
+        border-color: #ffb4a2;
+        box-shadow: 0 0 0 3px rgba(252, 213, 206, 0.4);
+        outline: none;
+      }
+    }
+  }
+
+  &__submit-button {
+    align-self: center;
+    padding: 0.85rem 2rem;
+    font-size: 1.125rem;
+    font-weight: 600;
+    color: #fff0eb;
+    background-color: #83494c;
+    border: none;
+    border-radius: 2rem;
+    font-family: inherit;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    box-shadow: 0 6px 20px rgba(131, 73, 76, 0.3);
+
+    &:hover {
+      background-color: #6a3a3d;
+      transform: scale(1.05);
+    }
+  }
+
+  &__login-link {
+    text-align: center;
+    margin-top: 2rem;
+    font-size: 1rem;
+
+    a {
+      color: #83494c;
+      font-weight: 600;
+      text-decoration: none;
+      transition: color 0.2s ease;
+
+      &:hover {
+        text-decoration: underline;
+        color: #5a2d2f;
+      }
+    }
+  }
+}
