+++ src/pages/AnalysisPage/AnalysisPage.jsx
@@ -3,11 +3,12 @@ import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import "./AnalysisPage.scss";
 import NavBar from "../../components/NavBar/NavBar";
-import photoIcon from "../../assets/icons/upload.png"; // Import your custom icon
+import ImageUpload from "../../components/ImageUpload/ImageUpload";
+import ScanResults from "../../components/ScanResults/ScanResults";
 
 function AnalysisPage() {
   const [image, setImage] = useState(null);
-  const [fileName, setFileName] = useState(""); // State for file name
+  const [fileName, setFileName] = useState("");
   const [results, setResults] = useState(null);
   const [error, setError] = useState("");
   const [isLoading, setIsLoading] = useState(false);
@@ -29,7 +30,7 @@ function AnalysisPage() {
     const file = e.target.files[0];
     if (file) {
       setImage(file);
-      setFileName(file.name); // Set the file name
+      setFileName(file.name);
     }
   };
 
@@ -84,44 +85,8 @@ function AnalysisPage() {
         <h1 className="analysis-page__title">Scan Your Product</h1>
         {error && <p className="analysis-page__error">{error}</p>}
 
-        {/* Image Upload */}
-        <div className="analysis-page__upload">
-          <div className="analysis-page__upload-group">
-            <div className="analysis-page__upload-icon">
-              <img
-                src={photoIcon}
-                alt="Upload Icon"
-                className="analysis-page__upload-icon-img"
-              />
-            </div>
+        <ImageUpload onImageUpload={handleImageUpload} fileName={fileName} />
 
-            <input
-              type="text"
-              readOnly
-              value={fileName}
-              placeholder="Upload Image"
-              className="analysis-page__upload-input"
-            />
-
-            <input
-              type="file"
-              id="image-upload"
-              accept="image/*"
-              onChange={handleImageUpload}
-              className="analysis-page__upload-hidden"
-            />
-
-            <button
-              type="button"
-              className="analysis-page__upload-browse"
-              onClick={() => document.getElementById("image-upload").click()}
-            >
-              Browse
-            </button>
-          </div>
-        </div>
-
-        {/* Image Preview */}
         {image && (
           <div
             className={`analysis-page__image-preview ${
@@ -136,7 +101,6 @@ function AnalysisPage() {
           </div>
         )}
 
-        {/* Scan Button */}
         <button
           onClick={handleScan}
           disabled={isLoading}
@@ -145,47 +109,7 @@ function AnalysisPage() {
           {isLoading ? "Scanning..." : "Scan"}
         </button>
 
-        {/* Results Display */}
-        {results && (
-          <div className="analysis-page__results" ref={resultsRef}>
-            <h2 className="analysis-page__results-title">Scan Results</h2>
-
-            {[
-              { label: "Beneficial Ingredients", data: results.beneficial },
-              {
-                label: "Potential Irritants",
-                data: results.potentialIrritants,
-              },
-              { label: "Harmful Ingredients", data: results.harmful },
-            ].map((section, i) => (
-              <div key={i} className="analysis-page__results-section">
-                <h3>{section.label}</h3>
-                {section.data && section.data.length > 0 ? (
-                  <div className="analysis-page__results-bubbles">
-                    {section.data.map((ingredient, index) => (
-                      <div
-                        key={index}
-                        className="analysis-page__bubble"
-                        style={{ animationDelay: `${index * 800}ms` }}
-                      >
-                        <div className="analysis-page__bubble-inner">
-                          <div className="analysis-page__bubble-front">
-                            {ingredient.name}
-                          </div>
-                          <div className="analysis-page__bubble-back">
-                            {ingredient.description}
-                          </div>
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                ) : (
-                  <p>No {section.label.toLowerCase()} detected.</p>
-                )}
-              </div>
-            ))}
-          </div>
-        )}
+        {results && <ScanResults results={results} resultsRef={resultsRef} />}
       </div>
     </div>
   );
