--- src/pages/ContactPage/ContactPage.scss
@@ -1,153 +1,0 @@
-@use "../../styles/partials/typography" as *;
-@use "../../styles/partials/variables" as *;
-
-.contact {
-  background: #fae1dd;
-  color: #83494c;
-  padding: 2rem;
-  font-family: "Garbata";
-
-  &__top {
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-    gap: 2rem;
-    margin-bottom: 2rem;
-  }
-
-  &__brand {
-    max-width: 300px;
-
-    &-title {
-      height: 6rem;
-    }
-
-    &-tagline {
-      font-size: 1rem;
-      font-weight: 300;
-      margin-bottom: 1.5rem;
-    }
-
-    &-info {
-      font-size: 0.9rem;
-      line-height: 1.6;
-      font-family: Helvetica;
-    }
-  }
-
-  &__links {
-    display: flex;
-    gap: 4rem;
-
-    &-column {
-      h3 {
-        font-size: 1.25rem;
-        font-weight: 600;
-        margin-bottom: 1rem;
-      }
-
-      ul {
-        list-style: none;
-        padding: 0;
-
-        li {
-          margin-bottom: 0.75rem;
-
-          a {
-            color: #6a3a3d;
-            text-decoration: none;
-            transition: color 0.3s ease;
-
-            &:hover {
-              color: #ffd7ba; // Apricot on hover
-            }
-          }
-        }
-      }
-    }
-  }
-
-  &__divider {
-    height: 1px;
-    background: #83494c;
-    margin: 2rem 0;
-    opacity: 0.3;
-  }
-
-  &__bottom {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    flex-wrap: wrap;
-    gap: 2rem;
-  }
-
-  &__social {
-    display: flex;
-    gap: 1rem;
-
-    &-icon {
-      width: 2rem; // Fixed width
-      height: 2rem; // Fixed height
-      color: #83494c; // Icon color
-      transition: color 0.3s ease, transform 0.3s ease;
-
-      &:hover {
-        color: #ffd7ba; // Apricot on hover
-        transform: translateY(-3px); // Lift effect on hover
-      }
-      &:first-child{
-        width: 2.5rem;
-        height: 2.5rem;
-      }
-    }
-  }
-
-  &__newsletter {
-    &-title {
-      font-size: 1.25rem;
-      font-weight: 600;
-      margin-bottom: 1rem;
-    }
-
-    &-form {
-      display: flex;
-      gap: 0.5rem;
-    }
-
-    &-input {
-      padding: 0.5rem;
-      border: 1px solid #83494c;
-      border-radius: 0.5rem;
-      font-family: "Garbata";
-      outline: none;
-
-      &::placeholder {
-        color: #6a3a3d;
-        opacity: 0.7;
-      }
-    }
-
-    &-button {
-      padding: 0.5rem 1rem;
-      background: #83494c;
-      color: #ece4db;
-      border: none;
-      border-radius: 0.5rem;
-      font-family: "Garbata";
-      font-weight: 600;
-      cursor: pointer;
-      transition: background 0.3s ease;
-
-      &:hover {
-        background: #6a3a3d;
-      }
-    }
-  }
-
-  &__copyright {
-    font-size: 0.9rem;
-    opacity: 0.8;
-    font-family: Helvetica;
-  }
-}
