--- src/pages/GlossaryPage/GlossaryPage.jsx
@@ -1,11 +1,0 @@
-import "./GlossaryPage.scss";
-
-function GlossaryPage() {
-  return (
-    <>
-      <h1>Glossary Page</h1>
-    </>
-  );
-}
-
-export default GlossaryPage;
