+++ src/pages/HomePage/HomePage.jsx
@@ -1,14 +1,12 @@
 import "./HomePage.scss";
+import { useState, useEffect } from "react";
 import FeatureCards from "../../components/FeatureCards/FeatureCards";
 import Footer from "../../components/Footer/Footer";
 import OurMission from "../../components/OurMission/OurMission";
-import { useState, useEffect } from "react";
 import BrandCarousel from "../../components/BrandCarousel/BrandCarousel";
 import HeroHeader from "../../components/HeroHeader/HeroHeader";
 
 function HomePage() {
-  const [isScrolled, setIsScrolled] = useState(false);
-  const [titleClass, setTitleClass] = useState("");
   const [isLoggedIn, setIsLoggedIn] = useState(false);
 
   useEffect(() => {
@@ -22,24 +20,6 @@ function HomePage() {
     localStorage.removeItem("token");
     setIsLoggedIn(false);
   };
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const scrollY = window.scrollY;
-      const viewportHeight = window.innerHeight;
-
-      if (scrollY > viewportHeight) {
-        setIsScrolled(true);
-        setTitleClass("in-nav");
-      } else {
-        setIsScrolled(false);
-        setTitleClass("");
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, []);
 
   return (
     <>
