+++ src/pages/QuizPage/QuizPage.scss
@@ -39,64 +39,6 @@
     letter-spacing: 0.5px;
   }
 
-  &__content,
-  &__determining,
-  &__result {
-    width: 100%;
-    text-align: center;
-    padding: 2rem;
-  }
-
-  &__question {
-    scroll-margin-top: 100px;
-    padding: 1rem;
-    border-radius: 1rem;
-    margin-bottom: 6rem;
-    animation: slideFadeIn 0.5s ease forwards;
-    opacity: 0;
-    transform: translateY(30px);
-
-    h3 {
-      font-size: 1.25rem;
-      font-weight: 600;
-      margin-bottom: 1rem;
-      text-align: center;
-    }
-  }
-
-  &__options {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    align-items: center;
-  }
-
-  &__option-button,
-  &__submit-button,
-  &__scan-button {
-    background-color: #83494c;
-    color: #fff0eb;
-    padding: 0.85rem 2rem;
-    border: none;
-    border-radius: 2rem;
-    font-family: inherit;
-    font-size: 1.125rem;
-    font-weight: 600;
-    cursor: pointer;
-    transition: all 0.3s ease;
-    box-shadow: 0 4px 12px rgba(131, 73, 76, 0.2);
-
-    &:hover {
-      background-color: #6a3a3d;
-      transform: scale(1.05);
-    }
-  }
-
-  &__submit-button {
-    margin-top: 1.5rem;
-    width: 100%;
-  }
-
   &__loading,
   &__error {
     text-align: center;
@@ -105,62 +47,7 @@
     color: #e63946;
     padding: 2rem 1rem;
   }
-  &__determining {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    min-height: 50vh; // Ensure it takes up enough space
-    opacity: 1;
-    transition: opacity 0.5s ease;
 
-    &--fade-out {
-      opacity: 0; // Fade out the loading state
-    }
-
-    p {
-      font-size: 1.25rem;
-      color: #83494c;
-      margin-top: 1rem;
-    }
-  }
-
-  &__spinner {
-    border: 4px solid #f3f3f3;
-    border-top: 4px solid #83494c;
-    border-radius: 50%;
-    width: 40px;
-    height: 40px;
-    animation: spin 1s linear infinite;
-    margin: 0 auto;
-  }
-
-  &__result {
-    opacity: 0;
-    transform: translateY(20px);
-    transition: opacity 0.5s ease 0.5s, transform 0.5s ease 0.5s; // Delay the transition
-
-    &--visible {
-      opacity: 1;
-      transform: translateY(0);
-    }
-
-    h2 {
-      font-size: 2rem;
-      font-weight: 600;
-      color: #83494c;
-      margin-bottom: 2rem;
-    }
-  }
-
-  @keyframes spin {
-    0% {
-      transform: rotate(0deg);
-    }
-    100% {
-      transform: rotate(360deg);
-    }
-  }
   @keyframes slideFadeIn {
     to {
       opacity: 1;
